Add more efficient multiprocessing
With this commit, more efficient multiprocessing is added to Mimir.
Previously, the entire `d` object was passed and serialized in the
multiprocessing pool. Now, only relevant components of `d` are isolated
and passed to the respecitve correctors, greatly improving memory
efficiency.

We still maintain an inefficient feature generation process for the sake
of being able to run Baran's correctors.